Types
General strictness about types, with some (minimal) casting

+++ include/logic.h
@@ -91,7 +91,7 @@ struct smt_world_anima_t {
 
   _Atomic(AnimaStatus) status;
 
-  _Atomic(int8_t) velocity;
+  _Atomic(uint32_t) velocity;
 };
 
 typedef struct smt_world_t SmtWorld;
